we have an interesting situation here
So when we try to make times methods same, we changed it to money to see
if it makes a difference. We made it compile and stuck on make the test work.
Wrote toString method to see what is wrong. Now it says we have 2 different
classess with same content. We should reconsider equals method for make it work